refactor: floating menu for operator
